+++ api/src/main/kotlin/com/sapuseven/untis/api/model/request/SchoolSearchParams.kt
@@ -8,6 +8,6 @@ import kotlinx.serialization.Serializable
 @ExperimentalSerializationApi
 data class SchoolSearchParams(
 	@EncodeDefault(EncodeDefault.Mode.NEVER) val search: String? = null,
-	@EncodeDefault(EncodeDefault.Mode.NEVER) val schoolid: Int = 0,
+	@EncodeDefault(EncodeDefault.Mode.NEVER) val schoolid: Long = 0,
 	@EncodeDefault(EncodeDefault.Mode.NEVER) val schoolname: String = ""
 ) : BaseParams()
